refactor(file-tree): split into FileTree/ directory
Pulls the tree-building helpers into tree.ts and gives the directory
header and file row their own components. The public surface
(./components/FileTree) is unchanged because the new folder exposes the
same FileTree export via index.tsx.
